requirements.txt
asgiref==3.7.2
axios==0.4.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
lxml==5.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.17.2
pytz==2023.3.post1
requests==2.31.0
rich==13.7.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
